+++ src/main/java/com/app/fitness/fitnesprogramapp/dtos/program/details/IntensityMetricDetailsDTO.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 public class IntensityMetricDetailsDTO {
     private Long id;
 
-    private int minimumIntensity;
-    private int maximumIntensity;
+    private double minimumIntensity;
+    private double maximumIntensity;
     private boolean isRange;
     private String title;
     private String metricSymbol;
